requirements.txt
streamlit
langchain_community
langchain
python-dotenv
langchain_core
langchain_groq
pypdf
langchain_huggingface
faiss-cpu